Enable address vectors to not be shared by endpoints
Some providers, such as EFA, do not allow address vectors to be shared
across endpoints. In general we want to share address vectors to reduce
overehead. However, if address vectors cannot be shared then an address
vector must be created for each endpoint. It's complicated somewhat by
scalable endpoints. In that case there are multiple transmit contexts
for the same endpoint, but the transmit contexts can share the same
address vector because there is only a single endpoint.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>