Use localThis once it is computed
PR #18323 noticed that 'var localThis = this.localize()'
was being called in several code paths and that as a result
it was unnecessary to compute it again.

However, it turns out that several method calls were made
on 'this' instead of 'localThis' which caused failures with
the tests utf8_values.chpl and utf8_misc.chpl.

This commit fixes the two failing cases to use the 'localThis'
that was computed rather than 'this'.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>