Stop disabling libnuma for hwloc builds
hwloc uses libnuma for the memory binding API so without libnuma hwloc
can't support operations like `hwloc_set_area_membind`. Some upcoming
improvements for 18286 and longer term Cray/chapel-private 1816 want to
use use the memory binding API, so I'm enabling autodetection of libnuma
here to see if there's any fallout from including it. We originally
disabled libnuma in 2921 to avoid some build failure on Crays, but this
was back when Cray machines defaulted to static linking and libnuma is
only available as a shared library. I suspect this won't cause any
issues now, but want to test it ahead of the PRs that will require it.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>