Add -O for knucleotide-strings
After PR #18323 we have been seeing this test timeout in some
configurations. It was already disabled in baseline and valgrind
configurations with PR #16302.

However, this test runs significantly (10x) faster if '-O' is provided in
order to request C/LLVM optimizations.  So, this PR just adds that in the
compopts.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>