Add initial support for HPE Apollo platforms
Add chplenv support for HPE Apollo (`CHPL_*_PLATFORM=hpe-apollo`). Like
for Cray CS, we default to gasnet-ibv and will use slurm launchers if
they are available. Also like Cray CS, there does not appear to be a way
to automatically detect Apollo, so users are required to set the
platform manually.

Part of Cray/chapel-private 1810

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>