Deprecate most locale task and thread counters.
We believe that locale.blockedTasks(), .queuedTasks(), .idleThreads(),
and .totalThreads() are unused.  Here, deprecate them.  We can back off
from this if we get feedback that any of them are needed.

The one test removed as part of this also tests locale.runningTasks(),
which we're not deprecating.  But that test was already notested because
other changes to task counting had rendered it obsolete, it doesn't test
.runningTasks() in a very deep way, and we have a number of other tests
that do test that specifically.  So it's okay to remove it entirely.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>