Fix clang++ 12 build by removing temporary char* variable
DavidL and Ahmad reported on Slack that PR #18329 (they think) broke
the compilation of driver.cpp, though this seems to be specific to
newer versions of clang++ (at least, it didn't show up with my default
version).  The error was:

driver.cpp:1514:32: error: object backing the pointer will be destroyed at the end of the
      full-expression [-Werror,-Wdangling-gsl]
            executableFilename:std::to_string(getpid()).c_str();

This fixes the issue by eliminating the temporary variable and
inserting the expression directly into the sprintf() that uses it.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>