Add method to print the deprecation message in documentation of symbols
Adds a new method to baseAST to go along with printDocsDescription.  This method
takes in the normal documentation for the symbol, the file to write to, the
number of tabs to use, and the deprecation message.  By calling it, we assume
the symbol being documented is deprecated.  It will then check the doc string
for anything about deprecation (if the documentation is not empty) and if there
not is already a mention, will print a warning about the symbol being
deprecated, using the provided deprecation message.

Also fix a typo while I'm here.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>