Remove migrate/strict flags from hwloc interleave call
Follow on to 18350. We only want to interleave new pages when faulting
them in, we don't want to move existing pages. This shouldn't impact
gasnet-ibv-large, which is what the interleaving code was added for, but
it's also called for ibv-fast where we want it to be a no-op. It didn't
seem to cause any issues before, but I think this is still the right
change. There was no inherent reason to have the migrate/strict flags
before, they were just copied from `chpl_topo_setMemLocalityByPages`.

While here also update our check to ensure that interleave is supported.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>