Run the LLVM Makefile on make
Since we build/install LLVM in stages, it can happen that
the LLVM build gets partway through and then fails or is aborted,
and then an incomplete install is left in the install directory.

The third-party/llvm/Makefile already checks for this case (by checking
for multiple files including the clang binary). However the
third-party/Makefile target for llvm was only checking for the existence
of the install directory.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>