Fix bug with implementation of `bigint.pow()` when the exponent is negative
The original implementation was calling `bigint.powm()` under the covers when
the exponent was negative, sending the `base` argument again as the `mod`
argument - however, through testing via Wolfram Alpha there were many cases
where sending the `base` as the `mod` value would result in a different result,
and actually calling the `bigint.mod` method in these cases would result in
segfaults.  Instead, since a negative exponent results in a fractional value for
all bases except 1 and -1, return 0 for most bases and return the base itself
when the base is 1 or -1.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>