Fix computePerfStats to find shortest regex
Previously, computePerfStats would search for performance keys
line-by-line. This caused us to find the wrong value if a key was a
substring of another one, but the longer key occurred first.

For example, if you had `non-regex time:` and `regex time:` keys, we
would find the `non-regex time:` value both times because `regex time:`
is in `non-regex time:`. Correct this by switching from searching line
at a time to a multiline regex, which will find the shortest match.

This was first noticed in some Arkouda benchmarks.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>