Extend the deprecation chpldoc tests to also lock in behavior with "no doc"
As expected and seen in the library documentation, the deprecation keyword and
generated message does not cause documentation to be generated when a symbol is
already marked with `pragma "no doc"`.

This change mostly impacts just the .chpl files (because "no doc" symbols don't
generate additional documentation) but there were a couple of cases where I
added a symbol that would influence the output - for classes to track that
"no doc" fields and methods in not deprecated classes are still hidden, and for
modules to verify that we don't generate a file for the "no doc" module.  The
latter case also required an update to the .catfiles to ensure we would notice
if we accidentally started generated files for those modules.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>