Fix an issue where this test wasn't completely looking in the right place
While adding "no doc" tests when the symbol is deprecated, I noticed that this
test of "no doc" failed to go looking for a nested module file - thankfully it
doesn't hide a bug today, but if we accidentally changed it so that nested
modules experienced a bug that top-level modules did not, this test as written
previously would not have noticed.  Update the .catfiles and the .good output
so that it will notice in the future.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>