Update compiler code to work with LLVM-12
Updating to LLVM-12 had some backward compatibility breaks that we hit. Update
the compiler so it will continue to work with both LLVM-11 and LLVM-12.

A new value in an enum required adding a new case to some switch statements.

A static method on DebugLoc became a static method on DILocation.

A static method on Module became a static method on StructType.

A method on VectorType got specialized into separate handling for
FixedVectorType and ScalableVectorType.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>