Fix `memThreshold` memory tracking optimization
18465 optimized memory leak tracking to avoid taking a lock when the real
allocation size is below `memThreshold`, but there were a few bugs in
the initial implementation.

For cases where the memory layer doesn't support `chpl_real_alloc_size`
(`CHPL_MEM=cstdlib`) we weren't handling the unknown size sentinel so we
weren't ever running the free hook, which made it look like we were
leaking all memory. Fix that to run the hook if the "real size" is 0
(the unknown sentinel.)

This was also broken under configurations that separately allocate
arrays (like `CHPL_COMM=ugni`). For those configs we were taking a comm
layer allocation and trying to ask jemalloc for the size of it, but this
results in mixed allocator calls, which is undefined behavior. Fix that
by making the higher level interface compute the size and passing it in
instead of computing it below the level where we know what allocator
memory came from.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>