If PE links with -dynamic, that also means dynamic linking.
In the PE command line driver's linking arguments, we have only been
recognizing '-Wl,-Bdynamic' as indicating that it will do dynamic
linking, because that's what the driver throws when it uses gcc to
link with.  But when it uses clang to link with, it instead throws
'-dynamic'.  Here, add recognition for that as well.  This fixes a
problem first seen with quite-new PrgEnv installs on an HPE Cray EX
lookalike system at a customer site.

While here, I also added curly braces in the immediate area of my
change around some single-statement code blocks that lacked them,
since that's our convention.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>