Speed up and reduce memory footprint of computePerfStats
18466 was intended to update computePerfStats to find the shortest regex
match. However, it didn't actually do this, and the new implementation
was memory intensive and causing OOMs on some machines. What we used to
do was just match for our regex line-by-line, which wasn't helpful when
we had some keys that were substrings of others. 18466 changed to match
on all the output, which helped match some substrings better but not all
and it made the regex match a slow memory hog for perf test that have a
lot of output (like many of the shootout benchmarks.)

Update computePerfStats here to switch from `match` to a simpler
`search`. This is faster and uses much less memory, but it gets us back
to finding the first key even if there's a shorter match later. To
combat that, first search for keys at the beginning of a line or
preceded by whitespace. This will help cases like:

```
non-regex time:
other regex time:
regex time:
```

It's not perfect, but some code I attempted with finditer to gather the
shortest match was just way too slow (30+ minutes for revcomp2.chpl). I
think this current approach will cover any likely substrings, but longer
term we should add an error message if we detect an ambiguity.

This also improves what we do in case of error in computePerfStats.
Previously, we might partially write to the .dat file and stdout, which
would leave the .dat file malformed and would cause any error messages
on stdout to be offset, preventing our sentinel error message matching
from working. Update to only print complete lines and only write to the
.dat file once we have all the values to write.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>