Add test locking in behavior Bryant discovered with "import M.{this}"
Tests a variant of the recently added tests of respecting the privacy of uses
and imports when validating the use/import lists, specifically around "this".

Note that this test still fails when the others worked.  This is because we did
not eliminate uses and imports behind private uses from our traversal (oops).
Looking into fixing that now

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>