Improve multilocale Regex
Previously Regex had some bugs when copying across locales for
init assignment. The subn method was also completely locale unaware.

We also now eagerly compile a local copy of the regex if the
assignment happens from a different locale, with the assumption that
uses of the regex will likely happen on that locale, which prevents
interrupting another task just to do a regex match.

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>