Adjust handling of the module list to search through for private uses/imports
The way the module list is created for scope resolution accidentally included
modules used or imported by modules that were privately used or imported.  For
instance, in the following use chain, if A were to be used by another module
Main, we would expect only A's contents to be visible to Main:

A
-> private use B
  -> use C

However, we were both including the use of B and including B's use of C, when
the use of B was private.  This change makes it so that we stop including a
chain when we see the root of it is private.

Unfortunately, we have to cut off exploration in two places.  We have to cut
it in the recursive case of `buildBreadthFirstUseImportList`, and we have to
recognize when our entry point into building the list would already be violating
that tactic.  The former will have a broader impact on scope resolution, the
latter will only matter for validating use and import lists (because if we're
looking at the contents of a use or import list, we know we're starting in the
middle of a chain).

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>