Remove old test directory for `chpldoc --devel`
Chpldoc used to support a `--devel` flag similar to that of the normal `chpl`
compiler, which would include arguments and methods created by the compiler
that would be known about by developers but not necessarily by users.  This
output stopped being useful when we adjusted the chpldoc pass to run right after
parsing, because the compiler generated symbols aren't inserted by that point.
some point, and we still might, but `chpldoc` and the compiler have changed so
much since that point that it doesn't seem reasonable to continue to maintain
these tests - we should just make new ones that reflect what changes when the
flag is thrown.
----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>