Allow comm=ofi debugging output to go to stderr.
Some platforms that use slurm buffer stdout so that the debugging output from comm=ofi
can be lost in a crash. By default the debugging output now goes to stderr. This can be changed
via the CHPL_RT_COMM_OFI_DEBUG_FNAME variable. Setting it to STDOUT causes the output to go to
stdout, and setting it to STDERR causes the output to go to stderr. If it's set to any other
value it is used as the prefix for output files, one per locale (this is the existing behavior).

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>