Add :throws: documentation to `subprocess.poll()` in the Spawn module
The method can throw under three conditions:
- The poll call was interrupted by a signal.  In this case, the user should
catch the error and make the call again, or adjust in some other way.
- The subprocess has been improperly set up so that the id being checked is not
a child process of the current process.  This would indicate a problem with the
implementation of the subprocess type - we'll throw a ChildProcessError, but I
don't think the user should be expected to handle such a case.
- The arguments sent to the underlying waitpid call were bad.  This would
indicate a problem with the implementation of poll, and I also don't think the
user can do anything about it.

I'm uncertain if we should document the ChildProcessError and EINVAL cases.  The
user can't really do anything to respond to it, they should just report to us
when they happen.  But I'm concerned that if we don't document these cases and a
user encounters one of them, they'll try to add a catch clause for it instead of
reporting it to us.  Maybe the method needs to be adjusted to only throw the
InterruptedError case and halt otherwise?

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>