Add tests for issue #12242
Add two tests for issue #12242 showing that we now have reasonable error
messages for this error case instead of an ICE.

The first issues errors about using 'this' before it is initialized. The second
comments out that use of 'this' and then has errors about default initializing
a generic channel.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>