Add documentation of potential errors thrown by `subprocess.wait()`
Note: not yet complete, as I haven't analyzed `subprocess.communicate()` which
is called by this method.

Also updated specific calls in the body to indicate potential errors thrown by
the call, and which ones are likely to be something the caller of `wait()` can
actually respond to (and thus are worth catching).

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>