Add a .future for issue #13159
Add a .future that has a user-defined class initializer and wants to give
a member type variable a default value at its declaration point.

It fails because we don't generate init methods when there are any
user-defined init methods, but we try to call one with an argument for each
generic field anyway.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>