Update these tests to use public uses/imports
Without a public use/import, these tests would fail on the limitation clause
listing a type that was not directly mentioned in the module.  But their purpose
is to ensure that the limitation clause impacts or doesn't impact an inherited
or forwarded method and to fully determine that, we need to get to the line that
tries to access the inherited or forwarded method.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>