Make CHPL_HOST_MEM & CHPL_TARGET_MEM configurable
These flags already existed and get passed through as their CHPL_MAKE
counterparts, but the function in chpl_mem.py was hardcoding it for
host.

CHPL_TARGET_{MEM,JEMALLOC} are new names for CHPL_{MEM,JEMALLOC} and
we keep the old name around for backwards compatibility, issuing a
warning if they are both set.

CHPL_HOST_{MEM,JEMALLOC} are separate flags, defaulting to cstdlib.

Currently, the only supported combination is linux with bundled (and
thus static), and mac with system (and thus shared)

When CHPL_HOST_MEM=jemalloc, we default CHPL_HOST_JEMALLOC to an
appropriate value if possible: bundled on linux and system on darwin

CHPL_HOST_MEM is used as part of the compiler build directory

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>