Add jemalloc host build
jemalloc needs different build configurations, like the "namespace"
prefix to be different between host and target builds, so they are
split into jemalloc/Makefile.{host,target}.include* files.

Previously, the jemalloc Makefile could assume it was always --target,
but now we pass that information down from the top level Makefile.

There are platform differences in overriding the system allocator --
eg. whether overriding malloc will override what strdup will use --
that makes it almost a requirement to use a dylib on mac. However,
there is an unresolved rpath issue preventing us from easily using a
bundled build of the shared library so we only allow system build on mac,
which works out because Homebrew provides this well.

On Linux, system packages of jemalloc are not as universal in what
they provide (eg. Fedora only provides libjemalloc.so.2 with no
jemalloc-config or pkgconfg .pc file) and it is possible to override
the system allocator by statically linking jemalloc, so we do.

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>