Adjust overload set checks to account for operators
Overload set checks would allow certain things for class methods that aren't
allowed otherwise - however, operator calls should be somewhat treated like
methods in this regard and the function was assuming they would have a method
token argument when that isn't always the case.  Adjust the check for operators,
and share code for checking the arguments so that it can be called on an
arbitrary argument as opposed to always the second one.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>