Add three tests of operators in import limitation clauses with {}
All work as expected with the changes to the parser (and the fix for the IO
operator).  Does not commit importing just the operator without {} - that still
generates a syntax error, there's another place I need to update.

Still TODO:
- fix for import without {}
- except lists testing (need individual tests for operators, especially since
they are defined in the same scope as the original type and so currently can be
found)

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>