Avoid type queries of LHS in ChapelHashtable._moveInit()
The copies of `_moveInit()` in ChapelHashtable.chpl have been doing
type queries on the lhs expression, both to make sure it matches the
type of the rhs expression (i.e., 'lhs.type == rhs.type') and to test
against 'nothing' (i.e., 'lhs.type == nothing').

This is problematic if the 'lhs' has a runtime type since the '.type'
queries try to compute its runtime type, but by definition the 'lhs'
in a _moveInit() is uninitialized.  When --dead-code-elimination was
on, the runtime type check never made it through to code generation,
but with --baseline, which turns it off, we'd try to query the lhs
runtime type, and end up reading uninitialized memory.

This PR refactors the moveInit() calls in order to:

* separate the nothing case out and rely on function resolution to
  resolve it rather than a type conditional

* remove the check that the LHS and RHS are of the same type, since
  this is a private function local to this module, and it seems
  certain that the types will match.

This works around the issue.  The one slightly unsatisfying thing is
that I don't have a good explanation for why the runtime type check is
unused in the generated --baseline code or what we could do to avoid
inserting it in the first place, since it doesn't seem used/useful.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>