Remove old chplenv logic about CHPL_COMM=ofi substrates
Remove some old and unused chplenv code for inferring substrates for
ofi. Substrates (providers under ofi) are selected at execution time and
not baked into the build time path, so this code isn't needed.

It was added and used in the initial OFI commit (d1aa83ea9b), but the
runtime code that used it has long since been removed.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>