Remove use of `in` by fixing `const ref` of extern
qio_regex_create_compile was missing a `const` for `options`. This
caused me to "fix" a compiler error by using the `in` intent for data,
causing a copy to be made.

This removes the `in` by fixing the `const` and reveals the
sub-optimal `localize` of the pattern. This will be addressed in the
near future in combination with other allocation/comm optimizations.

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>