Extend the ability to call functions from GPU kernels
Codegen for both GPU and CPU for functions that are called from a GPU kernel
or from functions that were called from other functions called from the kernel.

Add errors for accessing variables declared outside of the function for
functions called from the GPU. I think we'd like to allow this eventually,
but for the time being it is an error.

Add new tests for calling functions from functions, and testing the error about
accessing outer variables. Add .futures for recursive function calls and
calling 'halt' which currently each silently cause the code to run on the CPU.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>