Adjust codegenSizeofLLVM to use DataLayout
After PR #18542, we see valgrind errors / assert
failures (if LLVM is built with asserts on) for
these tests:

[Error matching program output for release/examples/benchmarks/hpcc/ra-atomics]
[Error matching program output for release/examples/benchmarks/hpcc/ra]
[Error matching program output for release/examples/benchmarks/hpcc/variants/ra-cleanloop]

The assertion error is coming from this line in SROA.cpp visitMemSetInst

  auto *C = cast<ConstantInt>(II.getLength());

Here, the way we were generating the sizeof for the memset instruction
was making a constant expression but not a constant int, causing the
failure.

This PR works around the problem by generating a constant int using
DataLayout.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>