Fix handling of non-zero exit status in mysystem()
This fixes an issue introduced in PR #18594 in which mysystem()
no longer generates a user error if the forked off process
returned a non-zero exit code.  An example impact of this is
that if 'make -f' failed, the compiler would generate a 0 exit
status and the test system would assume compilation succeeded
until it tried to run the resulting program, which wasn't there.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>