Improve multilocale regex performance
1) chpl__serialize for regex now calls chpl__serialize on the pattern
type (string/bytes) so that it gets passed inline in the buffer (when
small). this results in fewer comms

2) Because we are now by default going to have a local regex, we take
this to be the common case in eg _search_match. We take a copy of the
regex if its not local. Because matchign execution is local, we can
allocate some things on the stack with c_array. This results in less allocations

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>