Disable jemalloc merge/split hooks under ugni
We've been seeing sporadic errors under ugni that we've bisected to
18299, which enabled jemalloc merge/split hooks to reduce fragmentation
for configurations that use a fixed heap. It also enabled merge/split
for configs that use a dynamically extended heap, which today is just
ugni. For some reason this seems to be causing sporadic failures when
doing large BTE initiated GETs, so until we figure out the root cause
just disable merge/split for this config. This doesn't really impact
fragmentation since large arrays are separately allocated.

I view this as a short-term workaround for an upcoming point release,
but expect to spend more time investigating the root cause soon.

Part of https://github.com/Cray/chapel-private/issues/2678

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>