Add test showing that FormattedIO is no longer available by default
I'm adding the test from issue #18743, just to lock in the intended
behavior and make sure we don't backslide on this in the future.
My guess is that Lydia probably has added equivalent tests that
check for similar things in user code, but this was puzzling enough
(and the compile errors quickly enough) that it seemed beneficial
to add a test while I was looking at it.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>