Rename/add some helper functions and make sure to astr function name
A helper function I added to convert function names appropriately
was passed a UniqueString, and incorrectly returned the result of
`uniqueString.c_str()` without first calling `astr`. This is a
problem because small UniqueStrings are inlined (and thus stack
allocated). Call `astr` and make a comment to note what's going on.

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>