Only warn if CHPL_MEM and CHPL_TARGET_MEM disagree
This silences the warning that was causing a failed test in
test/compflags/link/sungeun/static_dynamic.skipif because sub_test.py
runSkipIf calls util/test/testEnv which takes the output from
printchplenv --all --internal --simple --no-tidy and sets them in the
environment. Because CHPL_MEM and CHPL_TARGET_MEM both get set by this
script, we were triggering the warning (which anything on stderr will
cause the skipif to fail).

I still think having the warning is a good thing and I think it's
reasonable to only warn when there is a mismatch.

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>